Deleted Destroyer classes, doing some Builder cleanup IBP-4283
