--- src/main/java/org/generationcp/middleware/operation/destroyer/Destroyer.java
@@ -1,37 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2012, All Rights Reserved.
- *
- * Generation Challenge Programme (GCP)
- *
- *
- * This software is licensed for use under the terms of the GNU General Public License (http://bit.ly/8Ztv8M) and the provisions of Part F
- * of the Generation Challenge Programme Amended Consortium Agreement (http://bit.ly/KQX1nL)
- *
- *******************************************************************************/
-
-package org.generationcp.middleware.operation.destroyer;
-
-import org.generationcp.middleware.hibernate.HibernateSessionProvider;
-import org.hibernate.Session;
-
-/**
- * Provides destroyer classes that can be used to delete logical/physical data in IBDBv2. Creates destroyer classes based on the given
- * local/central session parameters.
- *
- * @author Donald Barre
- */
-public abstract class Destroyer {
-
-	protected HibernateSessionProvider sessionProvider;
-
-	public Destroyer(HibernateSessionProvider sessionProvider) {
-		this.sessionProvider = sessionProvider;
-	}
-
-	public Session getActiveSession() {
-		if (this.sessionProvider != null) {
-			return this.sessionProvider.getSession();
-		}
-		return null;
-	}
-}
