--- src/main/java/org/generationcp/middleware/operation/destroyer/package.html
@@ -1,3 +1,0 @@
-<body>
-This package contains classes and methods that can be used to delete objects/entities from the IBDBv2 schema.
-</body>
